refactor: Line entity의 sections 분리
