+++ src/main/java/nextstep/subway/line/dto/LineResponse.java
@@ -30,7 +30,6 @@ public class LineResponse {
     }
 
     public static LineResponse of(Line line) {
-        line.sort();
         return new LineResponse(line.getId(), line.getName(), line.getColor()
                 ,line.getStations(), line.getCreatedDate(), line.getModifiedDate());
     }
